+++ CHANGELOG.md
@@ -5,8 +5,16 @@
   * Fix bug in teardown tool occuring when tearing dishes down from dish mode Operate
 * [AT-3272] 
   * change top level domain of DB host for archiver config
+<<<<<<< HEAD
 * [MAP-377]
   * Add OS shutdown command to gracefully shut down the talon boards prior to powering off the LRU
+=======
+* [AT-2992] 
+  * Fix bug collecting dish logs during scheduled pipeline runs
+* [AT-3261] 
+  * Add bitstream checksum based CBF firmware compatibility test
+  * Extend smoke test tooling for KAPB and pickup talon IPs automatically in KAPB
+>>>>>>> main
 
 ## 27.0.0
 * [AT-3260]